fix: CR and refactor out types and utils
