+++ packages/components/src/board-assets/tree-items/lanes/overlays-renderer.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import type { TreeOverlayProps } from '../../../tree/tree';
+import type { TreeOverlayProps } from '../../../tree';
 import { lanesContext } from '../../tree-items/lanes/lane-context';
 import type { TreeItemWithLaneData } from '../../tree-items/tree-item-with-lane-renderer';
 import { classes, st, vars } from '../../tree-items/lanes/overlays-renderer.st.css';
