+++ packages/components/src/board-assets/tree-items/tree-item-renderer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import type { TreeItemData } from '../';
 import { ChevronRightWixUiIcon } from '../../icons/wix-ui/chevron-right';
 import { SearchableText } from '../../searchable-text/searchable-text';
-import type { TreeItemProps } from '../../tree/tree';
+import { TreeItemProps } from '../../tree';
 import { classes, st, vars } from '../tree-items/tree-item-renderer.st.css';
 
 export const TreeItemRenderer: React.FC<TreeItemProps<TreeItemData>> = (props) => {
