+++ packages/components/src/board-assets/tree-items/tree-item-with-lane-renderer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { LaneData, calcLaneSize } from '../tree-items/lanes/lane-item-renderer';
 import { ElementData, ElementRenderer, calcElementSize } from '../tree-items/lanes/element-item-renderer';
 import { MarkerData, MarkerRenderer, calcMarkerSize } from '../tree-items/lanes/marker-item-renderer';
-import type { TreeItemInfo, TreeItemProps } from '../../tree/tree';
+import type { TreeItemInfo, TreeItemProps } from '../../tree';
 
 export type TreeItemWithLaneData =
     | LaneData<TreeItemWithLaneData>
