+++ packages/components/src/tree/index.ts
@@ -1,0 +1,3 @@
+export * from './tree';
+export * from './types';
+export * from './utils';
